requirements.txt
pandas
numpy
matplotlib
seaborn
google-cloud-bigquery
db-dtypes
jupyter